modify #10 プロジェクト構成を変更
