+++ src/functions/client/client-find.js
@@ -5,7 +5,7 @@ const dynamoDb = new AWS.DynamoDB.DocumentClient();
 
 module.exports = (event, callback) => {
   const params = {
-    TableName: 'todos',
+    TableName: 'Clients',
     Key: {
       id: event.pathParameters.id
     }
